chore: add sync for alias view
