+++ js/apps/system/_admin/aardvark/APP/react/src/views/views/searchAliasView/SearchAliasContext.tsx
@@ -10,6 +10,7 @@ import { mutate } from "swr";
 import { getApiRouteForCurrentDB } from "../../../utils/arangoClient";
 import { useIsAdminUser } from "../../../utils/usePermissions";
 import { ViewPropertiesType } from "./useFetchViewProperties";
+import { useSyncSearchViewUpdates } from "./useSyncSearchViewUpdates";
 import {
   getUpdatedIndexes,
   useUpdateAliasViewProperties
@@ -62,6 +63,8 @@ export const SearchAliasProvider = ({
     !!window.sessionStorage.getItem(`${initialView.name}-changed`)
   );
   const { onSave } = useUpdateAliasViewProperties({ setChanged });
+  useSyncSearchViewUpdates({ viewName: view.name });
+
   const onChange = (view: ViewPropertiesType) => {
     setView(view);
   };
