+++ js/apps/system/_admin/aardvark/APP/react/src/views/views/searchAliasView/useUpdateAliasViewProperties.ts
@@ -101,11 +101,6 @@ async function patchViewProperties({
         replace: true
       });
     }
-
-    window.arangoHelper.arangoNotification(
-      "Success",
-      `Updated View: ${view.name}`
-    );
   }
 }
